code review changes + blockID translation fix
